+++ src/main/java/org/jabref/gui/openoffice/OOBibBase.java
@@ -1,6 +1,5 @@
 package org.jabref.gui.openoffice;
 
-import java.io.IOException;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -26,7 +25,6 @@ import org.jabref.logic.openoffice.action.Update;
 import org.jabref.logic.openoffice.frontend.OOFrontend;
 import org.jabref.logic.openoffice.frontend.RangeForOverlapCheck;
 import org.jabref.logic.openoffice.oocsltext.CSLCitationOOAdapter;
-import org.jabref.logic.openoffice.oocsltext.CSLFormatUtils;
 import org.jabref.logic.openoffice.oocsltext.CSLUpdateBibliography;
 import org.jabref.logic.openoffice.style.JStyle;
 import org.jabref.logic.openoffice.style.OOStyle;
@@ -73,10 +71,10 @@ public class OOBibBase {
 
     private final OOBibBaseConnect connection;
 
+    private final OpenOfficePreferences openOfficePreferences;
+
     private CSLCitationOOAdapter cslCitationOOAdapter;
     private CSLUpdateBibliography cslUpdateBibliography;
-
-    private OpenOfficePreferences openOfficePreferences;
 
     public OOBibBase(Path loPath, DialogService dialogService, OpenOfficePreferences openOfficePreferences)
             throws
@@ -92,10 +90,8 @@ public class OOBibBase {
         if (cslCitationOOAdapter == null) {
             StateManager stateManager = Injector.instantiateModelOrService(StateManager.class);
             Supplier<List<BibDatabaseContext>> databasesSupplier = stateManager::getOpenDatabases;
-            OOStyle initialStyle = openOfficePreferences.getCurrentStyle(); // may be a jstyle, can still be used for detecting subsequent style changes in context of CSL
-            cslCitationOOAdapter = new CSLCitationOOAdapter(doc, databasesSupplier, initialStyle);
+            cslCitationOOAdapter = new CSLCitationOOAdapter(doc, databasesSupplier, openOfficePreferences);
             cslUpdateBibliography = new CSLUpdateBibliography();
-            CSLFormatUtils.setBibliographyProperties(openOfficePreferences);
         }
     }
 
@@ -614,7 +610,9 @@ public class OOBibBase {
                     }
 
                     // If "Automatically sync bibliography when inserting citations" is enabled
-                    syncOptions.ifPresent(options -> guiActionUpdateDocument(options.databases, citationStyle));
+                    if (citationStyle.hasBibliography()) {
+                        syncOptions.ifPresent(options -> guiActionUpdateDocument(options.databases, citationStyle));
+                    }
                 } finally {
                     // Release controller lock
                     doc.unlockControllers();
@@ -641,7 +639,6 @@ public class OOBibBase {
             OOError.from(ex).setTitle(errorTitle).showErrorDialog(dialogService);
         } catch (CreationException
                  | WrappedTargetException
-                 | IOException
                  | PropertyVetoException
                  | IllegalTypeException
                  | NotRemoveableException ex) {
@@ -900,8 +897,10 @@ public class OOBibBase {
                 OOError.fromMisc(ex).setTitle(errorTitle).showErrorDialog(dialogService);
             }
         } else if (style instanceof CitationStyle citationStyle) {
+            if (!citationStyle.hasBibliography()) {
+                return;
+            }
             try {
-
                 OOResult<XTextDocument, OOError> odoc = getXTextDocument();
                 if (testDialog(errorTitle, odoc.asVoidResult())) {
                     return;